requirements.txt
# Core ML and Data Processing
torch
transformers==4.36.2
peft==0.7.1
datasets==2.16.1
sentence-transformers==2.2.2
accelerate==0.25.0
huggingface_hub==0.19.4
# Database and Vector Store
pymongo==4.6.0
motor==3.3.2
chromadb==0.4.22

# Web Search Integration
requests==2.31.0
beautifulsoup4==4.12.2

# Streamlit Frontend
streamlit==1.29.0
plotly==5.17.0
pandas==2.1.4
numpy

# Financial Data
yfinance==0.2.28

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
nltk==3.8.1
regex